remove useless proxy elements and create new entity
