+++ src/Centreon/Domain/PlatformTopology/PlatformTopology.php
@@ -83,83 +83,6 @@ class PlatformTopology
     private $serverId;
 
     /**
-     * @var bool Indicate if the server is behind another server (n-1)
-     */
-    private $isLinkedToAnotherServer = false;
-
-    /**
-     * data retrieved from 'informations' table
-     * @var bool platform type
-     */
-    private $isRemote = false;
-
-    /**
-     * data retrieved from 'informations' table
-     * @var string|null central's address
-     */
-    private $authorizedMaster;
-
-    /**
-     * data retrieved from 'informations' table
-     * @var string|null
-     */
-    private $apiUsername;
-
-    /**
-     * data retrieved from 'informations' table
-     * @var string|null
-     */
-    private $apiCredentials;
-
-    /**
-     * data retrieved from 'informations' table
-     * @var string|null
-     */
-    private $apiScheme;
-
-    /**
-     * data retrieved from 'informations' table
-     * @var int|null
-     */
-    private $apiPort;
-
-    /**
-     * data retrieved from 'informations' table
-     * @var string|null
-     */
-    private $apiPath;
-
-    /**
-     * data retrieved from 'informations' table
-     * @var bool SSL peer validation activated
-     */
-    private $sslPeerValidationRequired = false;
-
-    /**
-     * data retrieved from 'options' table
-     * @var string|null
-     */
-    private $proxyUrl;
-
-    /**
-     * data retrieved from 'options' table
-     * @var int|null
-     */
-    private $proxyPort;
-
-    /**
-     * data retrieved from 'options' table
-     * @var string|null
-     */
-    private $proxyUsername;
-
-    /**
-     * data retrieved from 'options' table
-     * @var string|null
-     */
-    private $proxyCredentials;
-
-    /**
      * Validate address consistency
      *
      * @param string|null $address the address to be tested
@@ -185,57 +108,6 @@ class PlatformTopology
         }
 
         return $address;
-    }
-
-    /**
-     * @param int|null $port
-     * @return int
-     */
-    public function checkPortConsistency(?int $port): int
-    {
-        // checking
-        $port = filter_var($port, FILTER_VALIDATE_INT);
-        if (false === $port || 1 > $port || $port > 65536) {
-            throw new \InvalidArgumentException(
-                _("Central platform's data are not consistent. Please check the 'Remote Access' form")
-            );
-        }
-        return $port;
-    }
-
-    /**
-     * @param string|null $encryptedKey
-     * @return string|null
-     */
-    public function decryptApiCredentials(?string $encryptedKey): ?string
-    {
-        if (empty($encryptedKey)) {
-            return null;
-        }
-        // first key
-        $path = __DIR__ . "/../../../../";
-        require_once $path . "/src/Security/Encryption.php";
-        if (file_exists($path . '/.env.local.php')) {
-            $localEnv = @include $path . '/.env.local.php';
-        }
-
-        // second key
-        if (empty($localEnv) || !isset($localEnv['APP_SECRET'])) {
-            throw new \InvalidArgumentException(
-                _("Unable to find the encryption key. Please check the '.env.local.php' file")
-            );
-        }
-        $secondKey = base64_encode('api_remote_credentials');
-
-        try {
-            $centreonEncryption = new Encryption();
-            $centreonEncryption->setFirstKey($localEnv['APP_SECRET'])->setSecondKey($secondKey);
-            return $centreonEncryption->decrypt($encryptedKey);
-        } catch (\throwable $e) {
-            throw new \InvalidArgumentException(
-                _("Unable to decipher central's credentials. Please check the credentials in the 'Remote Access' form")
-            );
-        }
     }
 
     /**
@@ -389,267 +261,6 @@ class PlatformTopology
     public function setServerId(?int $serverId): self
     {
         $this->serverId = $serverId;
-        return $this;
-    }
-
-    /**
-     * @return bool
-     */
-    public function isLinkedToAnotherServer(): bool
-    {
-        return $this->isLinkedToAnotherServer;
-    }
-
-    /**
-     * @param bool $isLinked
-     * @return $this
-     */
-    public function setLinkedToAnotherServer(bool $isLinked): self
-    {
-        $this->isLinkedToAnotherServer = $isLinked;
-        return $this;
-    }
-
-    /**
-     * @return bool
-     */
-    public function getIsRemote(): bool
-    {
-        return $this->isRemote;
-    }
-
-    /**
-     * @param bool $type
-     * @return $this
-     */
-    public function setIsRemote(bool $type): self
-    {
-        $this->isRemote = $type;
-        return $this;
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getAuthorizedMaster(): ?string
-    {
-        return $this->authorizedMaster;
-    }
-
-    /**
-     * @param string $address
-     * @return $this
-     */
-    public function setAuthorizedMaster(string $address): self
-    {
-        $this->authorizedMaster = $address;
-        return $this;
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getApiUsername(): ?string
-    {
-        return $this->apiUsername;
-    }
-
-    /**
-     * @param string $username
-     * @return $this
-     */
-    public function setApiUsername(string $username): self
-    {
-        $this->apiUsername = $username;
-        return $this;
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getApiCredentials(): ?string
-    {
-        return $this->apiCredentials;
-    }
-
-    /**
-     * @param string|null $encryptedKey
-     * @return $this
-     */
-    public function setApiCredentials(?string $encryptedKey): self
-    {
-        $this->apiCredentials = $this->decryptApiCredentials($encryptedKey);
-        return $this;
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getApiScheme(): ?string
-    {
-        return $this->apiScheme;
-    }
-
-    /**
-     * @param string|null $schema
-     * @return $this
-     */
-    public function setApiScheme(?string $schema): self
-    {
-        $schema = trim($schema, '/');
-        $this->apiScheme = ($schema === 'https' ? 'https' : 'http');
-        return $this;
-    }
-
-    /**
-     * @return int|null
-     */
-    public function getApiPort(): ?int
-    {
-        return $this->apiPort;
-    }
-
-    /**
-     * @param int|null $port
-     * @return $this
-     */
-    public function setApiPort(?int $port): self
-    {
-        // auto resolving default scheme port
-        if (null === $port && null !== $this->apiScheme) {
-            if ('https' === $this->apiScheme) {
-                $this->apiPort = 443;
-                return $this;
-            }
-            if ('http' === $this->apiScheme) {
-                $this->apiPort = 80;
-                return $this;
-            }
-        }
-        $this->apiPort = $this->checkPortConsistency($port);
-        return $this;
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getApiPath(): ?string
-    {
-        return $this->apiPath;
-    }
-
-    /**
-     * @param string|null $path
-     * @return $this
-     */
-    public function setApiPath(?string $path): self
-    {
-        $path = trim(filter_var($path, FILTER_SANITIZE_STRING), '/');
-        if (empty($path)) {
-            throw new \InvalidArgumentException(
-                _("Central platform's data are not consistent. Please check the 'Remote Access' form")
-            );
-        }
-        $this->apiPath = $path;
-        return $this;
-    }
-
-    /**
-     * @return bool
-     */
-    public function isSslPeerValidationRequired(): bool
-    {
-        return $this->sslPeerValidationRequired;
-    }
-
-    /**
-     * @param bool $status
-     * @return $this
-     */
-    public function setSslPeerValidationRequired(bool $status): self
-    {
-        $this->sslPeerValidationRequired = $status;
-        return $this;
-    }
-
-    public function getProxyUrl(): ?string
-    {
-        return $this->proxyUrl;
-    }
-
-    /**
-     * @param string|null $url
-     * @return $this
-     */
-    public function setProxyUrl(?string $url): self
-    {
-        $path = filter_var($url, FILTER_SANITIZE_STRING);
-        if (empty($url)) {
-            throw new \InvalidArgumentException(
-                _("Central's platform path is not consistent. Please check the 'Remote Access' form")
-            );
-        }
-        $this->proxyUrl = $url;
-        return $this;
-    }
-
-    /**
-     * @return int|null
-     */
-    public function getProxyPort(): ?int
-    {
-        return $this->proxyPort;
-    }
-
-    /**
-     * @param int $port
-     * @return $this
-     */
-    public function setProxyPort(int $port): self
-    {
-        $this->proxyPort = $this->checkPortConsistency($port);
-        return $this;
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getProxyUsername(): ?string
-    {
-        return $this->proxyUsername;
-    }
-
-    /**
-     * @param string|null $username
-     * @return $this
-     */
-    public function setProxyUsername(?string $username): self
-    {
-        $username = filter_var($username, FILTER_SANITIZE_STRING);
-        if (empty($username)) {
-            throw new \InvalidArgumentException(
-                _("Central platform's data are not consistent. Please check the 'Remote Access' form")
-            );
-        }
-        $this->proxyUsername = $username;
-        return $this;
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getProxyCredentials(): ?string
-    {
-        return $this->proxyCredentials;
-    }
-
-    /**
-     * @param string|null $credential
-     * @return $this
-     */
-    public function setProxyCredentials(?string $credential): self
-    {
-        $this->proxyCredentials = $credential;
         return $this;
     }
 }
